requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
decorator==4.0.9
Django==1.9.5
django-celery==3.1.17
django-ipware==1.1.5
django-storages==1.4.1
geopy==1.11.0
ipython==4.2.0
ipython-genutils==0.1.0
kombu==3.0.35
oauthlib==1.0.3
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
python-twitter==2.2
pytz==2016.4
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
